Eliminate the need for the DEBUG_STAT preprocessor macro.
* configure.ac: We no longer need to #define DEBUG_STAT with
--enable-debug, since the relevant debug messages are controlled
instead with find's command-line option -D stat.
* find/defs.h (enum DebugOption): bitmask values are now expressed
as 1 << N instead of decimal values, simply to make it clearer
when we need to start worrying about which integer type is used to
hold the bitmask.
* find/parser.c (fallback_getfilecon): Print a debug message when
the -D stat option is in effect, whether or not find was
configured with --enable-debug.
(parse_version): DEBUG_STAT is no longer a (distinct) feature.