updatedb: use the common definitions of the bug-report URL etc.
* locate/Makefile.am (updatedb): Substitute PACKAGE_BUGREPORT,
PACKAGE_BUGREPORT_URL and PACKAGE_URL.  We do this in the Makefile
directly since updatedb.sh is not an output file generated
directly by configure (instead, it's generated by the Makefile
itself).
* configure.ac (PACKAGE_BUGREPORT_URL): AC_DEFINE defines only a
preprocessor macro.  In order to get a similar effect for
substituted output files, we need to use AC_SUBST.  Hence
introduce a shell variable $bugreport_url which we pass to both
AC_SUBST and AC_DEFINE_UNQUOTED (replacing the previous call to
AC_DEFINE).
* locate/updatedb.sh (usage): Use PACKAGE_URL and PACKAGE_BUGREPORT_URL.