fix bug #46784: frcode drops last char if no final newline
* locate/testsuite/config/unix.exp (locate_textonly): Don't append
a newline to the input file.  Instead use exactly the data
specified by the caller.  This means that in some cases the input
now lacks a final newline.  Ignore any warning message produced by
frcode.
* locate/frcode.c (main): If the final character of the line
returned by getdelim is not the delimiter, don't overwrite it with
a NUL character.  Instead, issue a warning message.
* NEWS: Mention this bugfix.