find: support list of file types for -type and -xtype
* find/defs.h (enum file_type): Add enumeration for all (supported)
file types.
(struct predicate): Replace 'type' by 'types' as an array of bool
for the above enum file_type.
* find/parser.c (insert_type): Treat the argument to -type and -xtype
as a comma-separated list of type letters: instead of storing the
single type of the predicate, save the searched type letters in the
above 'types' array.
* find/pred.c (pred_type): Change the comparison against the saved file
type predicate: now check if the type of the actual file is in the array
of searched 'types'.
* find/testsuite/test_type-list.sh: Add test.
* find/testsuite/Makefile.am: Reference the test.
* find/find.1: Document the new feature.
* doc/find.texi: Likewise.
* NEWS: Likewise.
RFE originally initiated by Young Mo Kang in
http://lists.gnu.org/archive/html/bug-findutils/2016-02/msg00025.html