doc: fix some -perm examples in find.info
* doc/find.texi (Mode Bits): Fix the description of the -perm examples
which search for the "022" mode bits: the match is for the file's group
and 'other' mode bits instead of for user and group.
While at it, remove a superfluous ';' in the adjacent example.
Bug introduced when adding the -perm examples in FINDUTILS-4.2.11.
* NEWS: Mention the fix.

Reported by Jacob Nevins <0jacobnk.gnu@chiark.greenend.org.uk> in
http://savannah.gnu.org/bugs/?50758