tests: skip sv-34079.sh properly if setting up the test fails
The test exited as a test failure (exit 1) also when set-up failed,
e.g. when running the test with TMPDIR on a file system with too
few free inodes.  Instead, the test should exit with 77 to skip the
test.  In other cases, exit properly with 99 to signal a test
framework failure.

* find/testsuite/sv-34079.sh (very_expensive_): Move the check for
RUN_VERY_EXPENSIVE_TESTS to the beginning, thus decreasing the
indentation level of the test code.  This is also a preparation
step for moving this check to a central place.
(skip_,framework_failure_): Define utility functions to skip the test
or terminate it as a test framework failure.
(get_ifree_): Define function to determine the number of free inodes.
(make_test_data): Use it, and skip early if we know that there are too
few free inodes on the current file system.
Use "seq -f" to build the file names, and directly feed them via xargs
to touch.  This avoids slower shell code.
Skip the test if touch fails; previously, the test exited as FAIL.
Move the creation of test files up, i.e., outside the if-condition,
and use trap to remove the temporary directory at exit.