doc: fix various syntax issues in find.1
Changes:
- Avoid warning "(macro BR): only 1 argument, but more are expected".
- Change misused SI (metric) numeric prefixes to the binary ones, like
a bare "1K blocks" to "1 KB blocks".
- Protect a full stop (.) with "\&", if it has a blank (white-space)
in front of or (ignoring transparent characters to the full stop) after
it, and it does not mean an end of a sentence.
- Change the name of a macro for two fonts (e.g., BR and IR) to one
letter, if there is only one argument.
- Use a macro to change to the italic font, instead of \fI [1], if
possible.
- Remove space in the first column, if not intented.
- Move a full stop (period) and a comma outside of a quoted text, if
it is at the end of the quote and does not end a quoted sentence.
- Change a HYPHEN-MINUS (code 0x55, 2D) to a minus (\-), if in front
of a name for an option.
- Split lines longer than 80 characters into two or more lines.
- Separate the sentences and subordinate clauses; each begins on a new
line. Or: Adjust space between sentences (two spaces).
- Fix the spelling of "kibi", "mebi", and "gibi".

* find/find.1: Do the above.

Patch contributed via https://sv.gnu.org/patch/?9585
Copyright-paperwork-exempt: Yes