find: fix -printf %Y output to 'N' for broken links
The format %Y shall output 'N' for broken links (ENOENT), 'L' for ELOOP,
and '?' for any other error when stat()ing the symlink target.

The %Y implementation implicitly fell back to lstat(); therefore it
output 'l' for dangling symlinks.

  $ ln -s ENOENT DANGLE
  $ find -D stat DANGLE -printf '%y %Y %p\n'
  fallback_stat(): stat(DANGLE) failed; falling back on lstat()
  l l DANGLE

* find/print.c (do_fprintf): For %Y, always follow links to determine
the type of the symlink target.
* find/testsuite/find.gnu/printf-symlink.exp: Add a test case for broken
links and the ELOOP case ...
* find/testsuite/find.gnu/printf-symlink.xo: ... with the expected output.
* NEWS (Bug fixes): Mention the fix.

Bug introduced in version v4.5.8 with commit '25e7c5fbf9'.