doc: clarify again exit status handling of -exec command {} +
A non-zero exit status of a command launched via the -exec predicate
family only leads to a non-zero exit status of 'find' itself when
the '+' form of the punctuation for the command is used.  This is
also what POSIX requires [1].
To put it the other way round: when the ';' form is used, the exit
status of 'find' is not affected.

[1] http://pubs.opengroup.org/onlinepubs/9699919799/utilities/find.html

* find/find.1 (-exec ... +): Explicitly mention that the exit status of
the launched command affects find's exit status when the '+' form is used.
To allow a future change, deliberately leave the behavior for failing
commands with the ';' form undocumented.
(-execdir): Likewise.
Previously, in the form committed with v4.6.0-63-gae424b95, that sentence
did not explicitly limit the behavior to the '+' form, and therefore
might have confused the user.

Fixes https://sv.gnu.org/bugs/?54509