doc: clarify find options -newer, -anewer, -cnewer
The previous description "File was modified more recently than file."
was hard to understand without markup, e.g. over the phone.

Furthermore, the description for symbolic links as reference file was
wrong in the man page for the -anewer and -cnewer options: in fact,
find(1) always uses the mtime of the reference file.

* doc/find.texi (Comparing Timestamps): Improve the wording for the
-anewer/-cnewer/-newer options.  Use 'reference' as the argument name.
Add the note from find.1 that the -H/-L options make find(1) use
the data modification time of the file being referenced.
* find/find.1 (-anewer,-cnewer,-newer): Improve the wording, and use
'reference' as argument name.
Fix the description for -anewer/-cnewer: when -H/-L is used with a
symbolic link as reference, then find always uses the mtime of that,
not atime or ctime.

Reported by Dan Jacobson <jidanni@jidanni.org> in
https://lists.gnu.org/r/bug-findutils/2019-04/msg00002.html