doc: add some previously-missing steps to the release instructions.
doc/find-maint.texi: point out that the version number in
configure.ac also needs to be updated.  Point out that if you did the
release update on a branch it needs to be merged and in any case that
these changes need to be pushed to origin/master.  Also prompt the
maintainer to check that the FTP upload actually worked.