doc: display version of the GNU findutils on the title page
The first page of the Texinfo documentation should display the package
version the documentation corresponds to.  Also change the title
from "Finding Files" to "GNU Findutils ..." to better reflect the
package name.  These changes align the documentation better with
that of other GNU packages like e.g. grep.

* doc/find.texi: Do the above.  Furthermore, display the VERSION also
in the first sentence after the @direntry list.
While at it, enclose the FDL licence text in a @quotation block.
* doc/find-maint.texi: Likewise.