find.1: improve formatting of EXAMPLES section
The EXAMPLES sesction was hard to read because it was not clear to
which sentence a sample command belonged.  Use '.SS' and '.IP' markup,
and also indent the sample commands nicer.
Furthermore, change the formatting of file names as italice consistently.
Avoid empty lines, by either explicitly marking as empty via '.', or
by adding the invisible zero-width character '\&'.

* find/find.1 (.SH EXAMPLES): Do the above.