doc: fix description of xargs options -i, --replace
The description of the --replace/-i options of xargs in the Reference
chapter said that the replace-str would only be allowed to be omitted
for -i while this holds also true for the long option --replace.

A similar description of the same options in Interspersing File Names
was missing the '[]' to mark the REPLACE-STR as optional for -i.

* doc/find.texi (xargs options): Document that omitting the REPLACE-STR
is allowed for --replace, too.  Further align the description with
that in ...
(Interspersing File Names): ... here.  Mark the replace-str as optional
for -i by surrounding with '[]'.