[doc] Use @var{name}, not NAME or @var{NAME} in "Hard Links".
The @var{name} renders as NAME in Info, to this inconsistency is hard
to spot there.  The inconsistency is easier to spot in the typeset
output (e.g. PDF, Postscript).

Our solution is to use @var{name} even inside @samp, since that
works (and is intended to, according to the documentation).

This fixes Savannah bug #65305.