maint: Add a convenience target for checking manual pages.
* configure.ac: New recursive target, check-all-manpages, for
convenience in lint-checking manual pages.
* find/Makefile.am: Add new target check-all-manpages-local.  This
checks the manpages only.  This is faster than invoking "make
check" so it is convenient when working with manual page updates.
* locate/Makefile.am: Likewise.
* xargs/Makefile.am: Likewise.
* po/Rules-findutils-no-ops: add dummy rules for the new
recursive and local targets. These need to succeed, but Automake
does not add these rules in po/ as it doesn't manage the Makefile
there.